review remarks #1
